Migrated EH-EXPORT to a separate crate.
