+++ crates/tool-eventhub-export/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "eh-export"
+version = "1.0.0"
+edition = "2021"
+authors = ["Breno RdV"]
+description = "EventHub Message Exporter Companion Tool."
+repository = "https://github.com/brenordv/rusted-toolbox"
+
+[dependencies]
+shared = { path = "../shared" }
+csv = "1.3.1"
+anyhow = "1.0.99"
+tokio = "1.47.1"
+serde_json = "1.0.142"
+clap = "4.5.45"
+tracing = "0.1.41"
+crossterm = "0.27.0"
+chrono = "0.4.41"
+sled = "0.34.7"
