+++ crates/tool-eventhub-export/src/main.rs
@@ -1,17 +1,21 @@
-use rusted_toolbox::tools::eh_export::cli_utils::{get_cli_arguments, print_runtime_info};
-use rusted_toolbox::tools::eh_export::eventhub_export_app::EventHubExporter;
-use rusted_toolbox::tools::eh_export::runtime_config_utils::{
-    apply_cli_overrides, validate_config,
-};
-use shared::constants::general::{EH_EXPORT_APP_NAME, EXIT_CODE_INTERRUPTED_BY_USER};
+use std::sync::Arc;
+use shared::constants::general::{EXIT_CODE_INTERRUPTED_BY_USER};
 use shared::eventhub::utils::config_utils::get_base_config_object;
 use shared::logging::app_logger::LogLevel;
 use shared::logging::logging_helpers::initialize_log;
 use shared::system::get_current_working_dir::get_current_working_dir;
 use shared::system::setup_graceful_shutdown::setup_graceful_shutdown;
 use shared::system::tool_exit_helpers::{exit_error, exit_success, exit_with_code};
-use std::sync::Arc;
+use crate::cli_utils::{get_cli_arguments, print_runtime_info};
 use tracing::{error, info};
+use crate::eventhub_export_app::EventHubExporter;
+use crate::runtime_config_utils::{apply_cli_overrides, validate_config};
+
+mod message_exporters;
+mod cli_utils;
+mod export_progress_tracker;
+mod eventhub_export_app;
+mod runtime_config_utils;
 
 /// Azure EventHub export tool - exports messages from local database to files.
 ///
@@ -51,7 +55,7 @@ use tracing::{error, info};
 /// handling possible user interruptions or errors during the process.
 #[tokio::main]
 async fn main() {
-    initialize_log(EH_EXPORT_APP_NAME, LogLevel::Info);
+    initialize_log(env!("CARGO_PKG_NAME"), LogLevel::Info);
 
     // Get CLI arguments
     let matches = get_cli_arguments();
