+++ crates/tools/Cargo.toml
@@ -34,10 +34,6 @@ path = "src/bin/split.rs"
 name = "eh-read"
 path = "src/bin/eh_read.rs"
 
-[[bin]]
-name = "eh-export"
-path = "src/bin/eh_export.rs"
-
 [dependencies]
 shared = { path = "../shared" }
 clap = { version = "4.5.45", features = ["derive"] }
