--- crates/tools/src/tools/eh_export/mod.rs
@@ -1,5 +1,0 @@
-pub mod cli_utils;
-pub mod eventhub_export_app;
-mod export_progress_tracker;
-pub mod message_exporters;
-pub mod runtime_config_utils;
